+++ tests/utils.py
@@ -24,7 +24,7 @@ from twisted.web.server import Request, Site
 from zope.interface import implementer
 
 from sydent.config import SydentConfig
-from sydent.sydent import Sydent, parse_config_dict
+from sydent.sydent import Sydent
 
 # Expires on Jan 11 2030 at 17:53:40 GMT
 FAKE_SERVER_CERT_PEM = """
@@ -70,14 +70,11 @@ def make_sydent(test_config={}):
 
     reactor = ResolvingMemoryReactorClock()
 
-    cfg = parse_config_dict(test_config)
-
     sydent_config = SydentConfig()
-    sydent_config.parse_from_config_parser(cfg)
+    sydent_config.parse_config_dict(test_config)
 
     return Sydent(
         reactor=reactor,
-        cfg=cfg,
         sydent_config=sydent_config,
         use_tls_for_federation=False,
     )
